+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.1.0 (2024-03-28)
 - **breaking:** removed `alloc` feature
-- **added:** fractal support for `SampleFn`
+- **breaking:** renamed `SampleFn` to `NoiseFn`
+- **added:** fractal support for `NoiseFn`
 
 ## 0.0.0 (2024-04-25)